+++ core/src/keyman_core.h
