+++ core/src/meson.build
@@ -59,6 +59,7 @@ kmx_files = files(
   'km_core_debug_api.cpp',
   'km_core_processevent_api.cpp',
   'jsonpp.cpp',
+  'util_normalize.cpp',
   'ldml/ldml_processor.cpp',
   'ldml/ldml_transforms.cpp',
   'ldml/ldml_markers.cpp',
